requirements.txt
fastapi==0.135.2
slowapi==0.1.9
uvicorn[standard]==0.29.0

# ── Database ───────────────────────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.29
# asyncpg==0.29.0  # Only needed for PostgreSQL (requires MSVC C++ build tools on Windows)
aiosqlite==0.20.0
alembic==1.13.1

# ── Pydantic & Settings ────────────────────────────────────────────────────────
pydantic>=2.9.0
pydantic-settings>=2.2.1
email-validator==2.1.1

# ── Authentication ─────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
bcrypt==4.1.3

# ── File Uploads ───────────────────────────────────────────────────────────────
python-multipart==0.0.9
aiofiles==23.2.1

# ── Cloud Storage (optional — only required when STORAGE_BACKEND=s3) ──────────
boto3>=1.34.0

# ── AI / LLM ───────────────────────────────────────────────────────────────────
openai==1.25.0
faiss-cpu==1.13.2
numpy>=1.26

# ── Document Parsing ──────────────────────────────────────────────────────────
pymupdf>=1.24.0
python-docx==1.1.0

# ── HTTP Client ───────────────────────────────────────────────────────────────
httpx==0.27.0

# ── Dev / Environment ─────────────────────────────────────────────────────────
python-dotenv==1.0.1
pytest==8.2.0
pytest-asyncio==0.23.6
